Handle data that is fetched on the same period as we publish
I have a data source that's updated every 60s, which is the same as my
publish interval.  This was a problem, because if the data was updated
between the last get and the publish, then we wouldn't see the data to
publish in this time interval.  When the next time interval started and
we did a get, we would consider that data too old.  As a result we'd
never publish it.

The fix is to do a get as part of publishing.  This leads to slightly
more data published than requested, but prevents gaps.